requirements.txt
tensorflow-gpu==1.12.0
tqdm==4.23.4
numpy==1.15.3